Изменил модель и апп

+++ commerce/shopping-store/src/main/java/ru/yandex/practicum/dto/ProductResponseDto.java
@@ -3,16 +3,15 @@ package ru.yandex.practicum.dto;
 import lombok.Data;
 import ru.yandex.practicum.enums.AvailabilityStatus;
 import ru.yandex.practicum.enums.ProductStatus;
-
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
-
 public class ProductResponseDto {
     private UUID id;
     private String name;
     private String description;
     private AvailabilityStatus availability;
     private ProductStatus status;
-    private Double price;
+    private BigDecimal price;
 }
